Refactor: Move ProjectItem into individual dir
