+++ src/components/ProjectItem/ProjectItem.scss
